`ssmmessages` and `ec2messages` are compliant when SSM is compliant (fixes #48)

+++ aws_allowlister/scrapers/tables/iso.py
@@ -3,15 +3,17 @@ from bs4 import BeautifulSoup, Tag
 from aws_allowlister.scrapers.aws_docs import get_aws_html
 from aws_allowlister.database.raw_scraping_data import RawScrapingData
 from aws_allowlister.shared.utils import clean_service_name
+from sqlalchemy.orm.session import Session
 
 
-def scrape_iso_table(db_session, link, destination_folder, file_name):
+def scrape_iso_table(db_session: Session, link: str, destination_folder: str, file_name: str, download: bool = True):
     html_file_path = os.path.join(destination_folder, file_name)
-    if os.path.exists(html_file_path):
-        os.remove(html_file_path)
 
     # get_aws_html gets the HTML from AWS docs and stores it locally, then returns the path
-    get_aws_html(link, html_file_path)
+    if download:
+        if os.path.exists(html_file_path):
+            os.remove(html_file_path)
+        get_aws_html(link, html_file_path)
 
     raw_scraping_data = RawScrapingData()
